Library/PackageCache/com.meta.xr.simulator@64.0.0/MetaXRSimulator/scripts/requirements.txt
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
fonttools==4.47.2
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.7.4
numpy==1.24.4
packaging==23.2
pillow==10.2.0
pixelmatch==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
typing==3.7.4.3
vrs==1.0.4
zipp==3.17.0
